cmd/vet: fix bug where incorrect exit code was returned
Fixes issue 8792.

This is a simple change that fixes the issue. It may be desireable
to opt for a larger code change that makes this problem less likely
to be inadvertedly reintroduced in the future. For instance, a vetMain()
func can be used similar to gofmtMain(), or the os.Exit call can be
deferred.

LGTM=r
R=golang-codereviews, r
CC=golang-codereviews
https://codereview.appspot.com/150850043

Committer: Rob Pike <r@golang.org>