dashboard: highlight positive changes Initially both positive and negative changes had solid background (green and red). Then we removed background of positive changes due to: https://code.google.com/p/go/issues/detail?id=8624 However, Ian noted:
"I did intuitively understand that + was an increase and - was a
decrease.  I did not notice how the numbers were segregated between
the third and fourth columns.  Now that you point it out, it is
obvious.  It would be more obvious if the numbers in the fourth column
were highlighted in green."

Give positive changes green *border* (not background),
so they are still visually distinguishable from negative even in grayscale.
This is especially beneficial for perf detail view,
because currently it looks like everything is either
negative or neutral in that view (only red and black).

LGTM=adg
R=adg
CC=golang-codereviews, iant, rsc
https://codereview.appspot.com/159970043